\!refresh now also updates roles messages

+++ commands/roles.js
@@ -2,9 +2,11 @@ const Discord = require("discord.js")
 
 module.exports.run = async (bot, message, args) => {
     var message_copy = message
-    try {
-        await message.delete()
-    } catch (error) { console.error(error) }
+    if (!args[0]) {
+        try {
+            await message.delete()
+        } catch (error) { console.error(error) }
+    }
     let roles = ""
     for (role in bot.roles[message_copy.guild.id]) {
         let emoji = bot.roles[message_copy.guild.id][role]
@@ -20,9 +22,13 @@ module.exports.run = async (bot, message, args) => {
         .setTitle(`🛡️ Current roles 🛡️`)
         .setColor(0xFF0000)
         .setDescription(roles)
-    try {
-        await message_copy.channel.send(embed)
-    } catch (error) { console.error(error) }
+    if (!args[0]) {
+        try {
+            await message_copy.channel.send(embed)
+        } catch (error) { console.error(error) }
+    } else {
+        return embed
+    }
 }
 
 module.exports.help = {
